Create sshattackblock_config.pm
This way, you can put your configuration in one place, and then update the blocker script as often as you need, without needing to patch each update of the blocker script. This is very useful when you have two different OSs to support, for example having logs in /var/log on one machine and /var/adm on the other, and you need to parse on different strings for each machine. Now you can change the blocker script, and no need to remember to change all these OS-specific bits in the bloack script, because they will be inherited from the PM file.